Downloads/recon_framework/requirements.txt
pymongo==4.12.0
python-dotenv==1.1.0
requests==2.32.3
discord-webhook==1.4.1
dnspython==2.7.0
